+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Watermark is hidden for screen readers
 - Moved asterisk-annotation on top of each form
 
+### Added
+
+- aria-hidden attribute to all Divider components
+
 ## [0.2.8] - 2021-12-23
 
 ### Changed